requirements.txt
Flask>=3.0.2
PyJWT>=2.8.0
flask-limiter>=3.8.0
bcrypt>=4.1.0
Flask-Login>=0.6.3
flask-cors>=4.0.0
peewee>=3.17.0
requests>=2.32.4
pyyaml
werkzeug
onvif-zeep>=0.2.12
wsdiscovery>=2.1.1
psycopg2-binary
opencv-python-headless>=4.9.0